requirements.txt
lxml==4.9.3
Requests==2.31.0
python-dotenv==1.0.0